requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
